launch script v1.3
it should properly check ship liqfuel resources now when checking for an LRB overlaping stage.
the flight plan is tweaked to work better with too powerful ships and also keep heavy ships in a nice flight. 
rendezvous are to come.
im thinking of maybe even adding user input but i'll have to dig further in to the input terminal. 
enjoy.
oh, default apoapsis is 130,000 meters. tweak it in the begininig of the script if you want another apoapsis.